Add send CAS and Learning Objectives as JSON feature
This commit introduces the feature of sending CAS types and Learning Objectives
as JSON files to the front when a get request is recieved by the server. As
a result, a new file directory has been introduced to support this feature, as
it contains all of the supporting csv files for this feature. With the
associated urls update. Tests have also been updated to reflect this feature.
Home page is updated to now show the avalible links to the developer to use.