Add security checks and issue tracker template
This commit introduces a script which will now check if the proper security
settings are in place as recommended in the Django docs.This check is also
added in circle.yml so every push will now run the security checks. The
string format bug is now fixed so that the linter will now include the pylintrc
file. Finally, security settings have now been updated to reflect a safer web
app.