Overwrite id fields for models
This commit introduces an overwrite of id fields for Activity, Entry, Student,
Adviser, and Coordinator. The new id will be a unique UUID field from which all
db queries should use from now on. Because id fields cannot be overwritten on
User models, the mk (model key) field will stand in for the id field for all
models that inherit from Person.